回転処理部分のテストを追加
